refactor(stores): migrate to composed campaign store, delete old monolithic store
- Deleted old monolithic campaign store (1,086 lines)
- Migrated create.tsx to use new composed store
- Updated storeExposure.ts to use new composed store
- Deleted dead page campaigns/[id].tsx (shadowed by [id]/index.tsx)
- Deleted old store tests (867 lines)
- Deleted dead page tests (1,047 lines)
- Deleted integration tests for old store (336 lines)
- Total: 3,562 lines deleted
- All 3,706 campaign tests pass
- Zero imports of old store remain
